+++ tests/models/test_case_small_network_with_ates/src/run_ates.py
@@ -5,8 +5,8 @@ from mesido.esdl.esdl_parser import ESDLFileParser
 from mesido.esdl.profile_parser import ProfileReaderFromFile
 from mesido.techno_economic_mixin import TechnoEconomicMixin
 from mesido.workflows.utils.adapt_profiles import (
-    adapt_hourly_profile_averages_timestep_size,
     adapt_hourly_year_profile_to_day_averaged_with_hourly_peak_day,
+    adapt_profile_to_averaged_timestep,
 )
 
 import numpy as np
@@ -159,7 +159,7 @@ class HeatProblem(
         """
         super().read()
 
-        adapt_hourly_profile_averages_timestep_size(self, 5 * 24)
+        adapt_profile_to_averaged_timestep(self, 5 * 24)
 
 
 class HeatProblemPlacingOverTime(HeatProblem):